src/requirements.txt
sympy
pwn
python-dotenv